requirements.txt
fastapi[all]
sqlalchemy
rich
pyjwt
passlib
bcrypt
omegaconf
hydra-core
llama-cpp-python
aiosqlite
